Partially update to 1.20.1
