+++ src/main/java/xyz/amymialee/mialeemisc/MialeeMiscClient.java
@@ -14,7 +14,6 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.registry.Registries;
 import net.minecraft.resource.ResourceType;
 import net.minecraft.util.Identifier;
-import xyz.amymialee.mialeemisc.client.InventoryItemRenderer;
 import xyz.amymialee.mialeemisc.cooldowns.IdentifierCooldownHolder;
 
 import java.util.Set;
@@ -38,17 +37,5 @@ public class MialeeMiscClient implements ClientModInitializer {
                 }
             });
         });
-    }
-
-    public static void registerInventoryItem(Item item) {
-        var itemId = Registries.ITEM.getId(item);
-        var inventoryItemRenderer = new InventoryItemRenderer(itemId);
-        ResourceManagerHelper.get(ResourceType.CLIENT_RESOURCES).registerReloadListener(inventoryItemRenderer);
-        BuiltinItemRendererRegistry.INSTANCE.register(item, inventoryItemRenderer);
-        ModelLoadingRegistry.INSTANCE.registerModelProvider((manager, out) -> {
-            out.accept(new ModelIdentifier(itemId, "inventory"));
-            out.accept(new ModelIdentifier(new Identifier(itemId.getNamespace(), itemId.getPath() + "_handheld"), "inventory"));
-        });
-        INVENTORY_ITEMS.add(item);
     }
 }
