+++ src/main/java/xyz/amymialee/mialeemisc/mixin/client/ItemRendererMixin.java
@@ -1,11 +1,5 @@
 package xyz.amymialee.mialeemisc.mixin.client;
 
-import com.mojang.blaze3d.systems.RenderSystem;
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.font.TextRenderer;
-import net.minecraft.client.network.ClientPlayerEntity;
-import net.minecraft.client.render.BufferBuilder;
-import net.minecraft.client.render.Tessellator;
 import net.minecraft.client.render.item.ItemModels;
 import net.minecraft.client.render.item.ItemRenderer;
 import net.minecraft.client.render.model.BakedModel;
@@ -14,53 +8,19 @@ import net.minecraft.client.world.ClientWorld;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
-import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.World;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import xyz.amymialee.mialeemisc.MialeeMiscClient;
-import xyz.amymialee.mialeemisc.cooldowns.IdentifierCooldownManager;
-import xyz.amymialee.mialeemisc.items.ICustomCooldownsItem;
 
 @Mixin(ItemRenderer.class)
 public abstract class ItemRendererMixin {
-    @Shadow protected abstract void renderGuiQuad(BufferBuilder buffer, int x, int y, int width, int height, int red, int green, int blue, int alpha);
+
     @Shadow @Final private ItemModels models;
-
-    @Inject(method = "renderGuiItemOverlay(Lnet/minecraft/client/font/TextRenderer;Lnet/minecraft/item/ItemStack;IILjava/lang/String;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isItemBarVisible()Z"))
-    public void mialeeMisc$cooldownOverlay(TextRenderer renderer, ItemStack stack, int x, int y, String countLabel, CallbackInfo ci) {
-        var clientPlayerEntity = MinecraftClient.getInstance().player;
-        if (clientPlayerEntity != null && stack.getItem() instanceof ICustomCooldownsItem item) {
-            var array = item.mialeeMisc$getCooldownIdentifiers();
-            for (var i = 0; i < array.length; i++) {
-                var progress = IdentifierCooldownManager.get(clientPlayerEntity).getCooldownProgress(array[i], MinecraftClient.getInstance().getTickDelta());
-                if (progress > 0.0f) {
-                    RenderSystem.disableDepthTest();
-                    RenderSystem.disableTexture();
-                    RenderSystem.enableBlend();
-                    RenderSystem.defaultBlendFunc();
-                    var tessellator2 = Tessellator.getInstance();
-                    var bufferBuilder2 = tessellator2.getBuffer();
-                    this.renderGuiQuad(bufferBuilder2, x + MathHelper.ceil(16f / array.length) * i, y + MathHelper.floor(16.0f * (1.0f - progress)), MathHelper.ceil(16f / array.length), MathHelper.ceil(16.0f * progress), 255, 255, 255, 127);
-                    RenderSystem.enableTexture();
-                    RenderSystem.enableDepthTest();
-                }
-            }
-        }
-    }
-
-    @Inject(method = "renderGuiItemOverlay(Lnet/minecraft/client/font/TextRenderer;Lnet/minecraft/item/ItemStack;IILjava/lang/String;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/player/ItemCooldownManager;getCooldownProgress(Lnet/minecraft/item/Item;F)F"), cancellable = true)
-    public void mialeeMisc$cooldownStopper(TextRenderer renderer, ItemStack stack, int x, int y, String countLabel, CallbackInfo ci) {
-        var clientPlayerEntity = MinecraftClient.getInstance().player;
-        if (clientPlayerEntity != null && stack.getItem() instanceof ICustomCooldownsItem) {
-            ci.cancel();
-        }
-    }
 
     @Inject(method = "getModel", at = @At("HEAD"), cancellable = true)
     private void mialeeMisc$heldItemModel(ItemStack stack, World world, LivingEntity entity, int seed, CallbackInfoReturnable<BakedModel> cir) {
