+++ src/main/resources/mialeemisc.mixins.json
@@ -11,27 +11,16 @@
     "ItemStackMixin",
     "LivingEntityMixin",
     "PlayerEntityMixin",
-    "ProjectileUtilMixin",
     "RecipeManagerMixin",
     "ServerPlayerEntityMixin",
     "accessors.LlamaEntityAccessor",
     "accessors.ProjectileAttackGoalAccessor",
-    "goals.BowAttackGoalMixin",
-    "goals.CrossbowAttackGoalMixin",
-    "goals.LlamaProjectileAttackGoalMixin",
-    "goals.ProjectileAttackGoalMixin",
-    "goals.TridentAttackGoalMixin",
-    "goals.WitchProjectileAttackGoalMixin",
-    "mobs.AbstractSkeletonEntityMixin",
-    "mobs.LLamaEntityMixin",
-    "mobs.MobEntityMixin",
-    "mobs.PiglinEntityMixin",
-    "mobs.WitchEntityMixin",
-    "mobs.WitherEntityMixin"
+    "mobs.MobEntityMixin"
   ],
   "client": [
     "client.ClientPlayerEntityMixin",
     "client.CreativeInventoryScreenMixin",
+    "client.DrawContextMixin",
     "client.EntityRenderDispatcherMixin",
     "client.ItemRendererMixin",
     "client.MinecraftClientMixin",
